chore: implement UserDetails to work with custom UserDetailsService
Took 2 minutes
